Intro and humanVshuman completed
The introduction of two question(player choose(X or O) and opponent (pc or human)) was done. 
At same time the playersVsplayer mode is ready, which start over again, after 2 secodnds,  when the game is over. 
In addition there are a couple of functions that will be helpful in the development of the AI